Move logsdb tests to logsdb plugin (#114952)
